requirements.txt
google-genai
fastapi
uvicorn[standard]
python-multipart
aiofiles
Jinja2
google-api-python-client
youtube-transcript-api
tqdm